requirements.txt
fastapi~=0.110.0
uvicorn[standard]~=0.30.0
sqlalchemy~=2.0.0
psycopg2-binary~=2.9.9
geoalchemy2~=0.14.0
shapely~=2.0.0
python-dotenv~=1.0.0
python-jose~=3.3.0
passlib[bcrypt]~=1.7.4
pydantic~=1.10.0

